data-processor/src/requirements.txt
pika==1.2.1
python-dateutil
nltk
sklearn
psycopg2-binary